feat(global): настроить single-spa
Jira issue: VEGA-415
