--- src/vega-root-config.ts
@@ -1,21 +1,0 @@
-import { registerApplication, start } from "single-spa";
-import {
-  constructApplications,
-  constructRoutes,
-  constructLayoutEngine,
-} from "single-spa-layout";
-
-const routes = constructRoutes(
-  document.querySelector("#single-spa-layout") as HTMLTemplateElement
-);
-const applications = constructApplications({
-  routes,
-  loadApp({ name }) {
-    return System.import(name);
-  },
-});
-const layoutEngine = constructLayoutEngine({ routes, applications });
-
-applications.forEach(registerApplication);
-layoutEngine.activate();
-start();
